\name{SeedVigorIndex}
\alias{seed_VI}
\title{Seed Vigor Index}
\usage{
seed_VI(seedling_length, germ_percent)
}
\arguments{
\item{seedling_length}{seedling length (in cm)}
\item{germ_percent}{germination percentage}
}
\value{
Vigor Index (VI)
}
\description{
Measurement of Seed Vigor Index using seedling length and germination percentage.
}
\examples{
seedling_length <- 7.2
germ_percent <- 52
seed_VI(seedling_length, germ_percent)
}
\references{
Ling L, Jiafeng J, Jiangang L, Minchong S, Xin H, Hanliang S, Yuanhua D (2014) Effects of cold plasma treatment on seed germination and seedling growth of soybean. Sci Rep 4:5859.\\
Dutta, S. K., Layek, J., Akoijam, R. S., Boopathi, T., Vanlalhmangaiha, Saha, S., & Prakash, N. (2019). Seaweed extract as natural priming agent for augmenting seed quality traits and yield in Capsicum frutescens L. Journal of applied phycology, 31, 3803-3813.
}
\keyword{Seed Vigor Index}

